requirements.txt
CairoSVG==2.7.1
datasets==3.0.2
groq==0.22.0
huggingface_hub==0.27.1
instructor==1.7.8
loguru==0.7.3
lpips==0.1.4
numpy==2.2.4
openai==1.70.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.3
pdf2image==1.17.0
Pillow==11.1.0
pydantic==2.11.1
pytorch_msssim==1.0.0
sacrebleu==2.4.3
TexSoup==0.3.1
TexSoup==0.3.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.66.6
zss==1.2.0
